requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
Cython==0.29.37
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.0
fsspec==2025.9.0
gevent==25.9.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
polars==1.33.1
psutil==7.1.0
pynmea2==1.19.0
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==6.0
zope.interface==8.0
